+++ src/modules/home/index.ts
@@ -1,0 +1,2 @@
+export * from './ManualSearchSection'
+export * from './WelcomeMessageSection'
