+++ src/pages/index.tsx
@@ -10,9 +10,7 @@ import { PageContent } from '../components/PageContent'
 import { SiteHeader } from '../components/SiteHeader'
 import { WebsiteEntry } from '../components/WebsiteEntry'
 import { fetchTrustPositif } from '../functions/fetchTrustPositif'
-import { ManualSearchSection } from '../modules'
-import { ExplanationSection, WhatIsThisSection } from '../modules/about'
-import { WelcomeMessageSection } from '../modules/home/WelcomeMessageSection'
+import { ManualSearchSection, WelcomeMessageSection } from '../modules/home'
 import type { PSEData } from '../types/PSEData'
 import { generateBlockList } from './api/fetchBlocked'
